refactoring code with Martin's reviews
