+++ bop_toolkit_lib/call_vsd_worker.py
@@ -3,7 +3,7 @@ import os
 import time
 
 from bop_toolkit_lib.renderer import create_renderer
-from bop_toolkit_lib.renderer_batch import POSE_ERROR_VSD_ARGS
+from bop_toolkit_lib.pose_error import POSE_ERROR_VSD_ARGS
 from bop_toolkit_lib import inout, misc, pose_error
