+++ bop_toolkit_lib/pose_error.py
@@ -14,6 +14,57 @@ from bop_toolkit_lib import misc
 from bop_toolkit_lib import visibility
 
 
+# Standard Library
+from typing import Optional
+from dataclasses import dataclass
+
+
+@dataclass
+class POSE_ERROR_VSD_ARGS:
+    # all args in pose_error.vsd
+    R_e: Optional[np.ndarray] = None
+    t_e: Optional[np.ndarray] = None
+    R_g: Optional[np.ndarray] = None
+    t_g: Optional[np.ndarray] = None
+    depth_im: Optional[np.ndarray] = None
+    K: Optional[np.ndarray] = None
+    vsd_deltas: Optional[float] = None
+    vsd_taus: Optional[list] = None
+    vsd_normalized_by_diameter: Optional[bool] = None
+    diameter: Optional[float] = None
+    obj_id: Optional[int] = None
+    step: Optional[str] = None
+
+    def from_dict(self, data):
+        for key, value in data.items():
+            setattr(self, key, value)
+        return self
+
+    def to_file(self, path):
+        for key, value in self.__dict__.items():
+            if value is None:
+                raise ValueError("Field {} is None".format(key))
+            value = np.array(value)
+        np.savez(path, self.__dict__)
+
+    def from_file(path):
+        args = POSE_ERROR_VSD_ARGS()
+        data = np.load(path, allow_pickle=True)
+        data = data["arr_0"].item()
+        for key, value in data.items():
+            if key == "vsd_taus":
+                setattr(args, key, list(value))
+            if key == "vsd_normalized_by_diameter":
+                setattr(args, key, bool(value))
+            if key == "step":
+                setattr(args, key, str(value))
+            if key == "obj_id":
+                setattr(args, key, int(value))
+            else:
+                setattr(args, key, value)
+        return args
+
+
 def vsd(
     R_est,
     t_est,
